MLE-28334 Copilot Suggested Fixes
+ AbstractFromSearchFragmentTest.java: Refactored FromSearchDocsWithFragmentTest.java and FromSearchWithFragmentTest.java
+ Changed JavaDoc for fromsearchDocs in PlanBuilderBase.java
